requirements.txt
google-generativeai
pandas
openpyxl
python-dotenv
Pillow
openai
langchain
langchain-google-genai
langchain-openai
langchain-core
